Fetch content items before routing and add a fall-through route
Before routing, try to fetch the content item from the content store.
We fetch the content item in the controller anyway, but by fetching
it before routing, we can make routing decisions based on the content
item, such as checking whether it’s a taxon or guidance content.

Armed with this advanced routing, we can also add a “fall-through”
route, which takes any paths that are neither taxons nor guidance
content, and simply fetches the full HTML from gov.uk and spits it out
verbatim.

This covers cases such as the gov.uk homepage, and any non-guidance
content where we don’t want to display navigation.