Fetch most popular content for the Blue Box
This commit fetches up to 7 content items, sorted by popularity
(descending) and then sorted by title from the search API.

Those links are then displayed in accordion and leaf pages only.

For the purposes of the next round of research, we are actively removing
2 pages from those lists, because we want to see if people find them via
the other navigation elements. Those pages are:

- /teacher-training-funding
- /nhs-bursaries