Hardcode Parenting and Childcare taxonomy
The Parenting and Childcare taxonomy currently exists in a draft format
on the Draft Content Store. We would like to test the navigation
prototype with the draft taxonomy, but cannot access the draft content
store unless we deploy the prototype to GOV.UK infrastructure.

Since deploying to GOV.UK infrastructure would make the deployment
process harder, we instead hardcode the draft taxonomy file in a JSON
file retrieved with a Rake task from Content Tagger.

When we make a request for a content item, we first check if this
content item exists in the local JSON file. If it does, then return
that instead.

This change also updates the GOV.UK homepage to point to the correct
taxon page, which doesn’t exist as a Production feature yet.