Adapt SchemaFinderService for content_item
SchemaFinderService now returns both page_schema and content_item.

The page_schema method uses the JSON structure proposed in
f5b5945c3c7942104769bcd0ec9a5e0e14462c61 and parses it into an
OpenStruct object for ease of access in a ruby like style.

This commit also introduces GovukNavigationHelpers::NavigationHelper which accepts a content_item.

The reason is to be able to use the breadcrumbs and related_items from the NavigationHelper,
which requires a content_item and handles both breadcrumbs and related_items structure.

Trello:
https://trello.com/c/z5tqVWsi/8-implement-example-in-govuk-services-prototype-app