Add new task based sidebar navigation
This adds sidebar navigation listing all the tasks in the task list and
highlighting the one that the current page sits within.

The navigation is only shown on pages that are contained in the navigation
steps.  It does this by checking the base path of the current page against
those that are within the items listed in "config/task_nav/how-to-become-a-childminder.json"

At present "forms" and organisation pages are not included in this navigation.