Add links to the tasklist page to browse
We want links to our tasklist page to appear in the taxon page for childcare.

We replace the last link in the "Most viewed in this topic" section so we
maintain 5 links.  Anymore becomes a bit overwhelming.

The other link ends up eighth in the "Becoming a childcare provider" accordion
section.