Insert a hard coded element as the top result
We want to add the learn to drive page as the top result.  We envisage that
any lab search will be roughly in this topic, so it's not too bad.

I've used a direct route as opposed to a constraint (like the others) as this
will only ever apply to the search path.