Use content item base path and prefixes for nav
We want to show the nav on all parts of a relevant guide.  To do this we use
the base_path of the json content_item to determine if the page needs the nav
shown.

The content item base_path is the same as the url path in the majority of
cases except in guides and other multipage formats.

In the example of the guide page:
https://www.gov.uk/theory-test/reading-difficulty-disability-or-health-condition

The URL path is:
/theory-test/reading-difficulty-disability-or-health-condition

The content item base path is:
/theory-test

We see if any of the paths in our task config start with the content item
base path so that it shows on pages like: https://www.gov.uk/theory-test/pass-mark-and-result
which is another part in the same guide (and also has the same base_path as
above)

We also want to show it on other relevant content which is why we've added
another section to the task_nav configuration links section.