Fix rendering of Browse pages through AJAX request
This route needs to be removed as the whole logic depends on finding
a browse page by the `params[:top_level_slug]`. What this was doing was
matching a file.json from the `params[:base_path]`, by doing so there
was no `params[:top_level_slug]` so the `/browse` pages where not
working under AJAX request.