Fall through to GOV.UK for most things
We set the cookies for the headers and sidebars for get-a-divorce and
end-civil-partnership content to be shown.

At this point we just want to show what's deployed on GOV.UK so we just fall
through to proxying.

We also redirect requests to the task list pages so that the navigation links
work OK.