Add sticky header to the prototype[WIP]
A script and css link has been added to the page so that the sticky header functionality can be added (paths are to be updated).

In the task list header, The `track-click` module is currently set as the `data-module` attribute. We want to add `tasklistheader` to this but it is likely `data-module` only accepts one value. We will replace `track-click` for now, as it can be added in later once the sticky header code has been implemented.

We also changed the routes so that content pages  will now go to the new action that deals with replacing the data attribute.